+++ src/components/Button/index.ts
@@ -1,2 +1,4 @@
 export { BaseButton } from "./BaseButton";
+export { CartActionButton } from "./CartActionButton";
+export { CartButton } from "./CartButton";
 export { ShopButton } from "./ShopButton";
